% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{griss_melanoma_result}
\alias{griss_melanoma_result}
\title{Example Camera result created based on the
melanoma induction study by Griss et al.}
\format{An object of class \code{ReactomeAnalysisResult} of length 1.}
\usage{
griss_melanoma_result
}
\description{
The result is stored as a \code{\link[ReactomeGSA]{ReactomeAnalysisResult}} object.
}
\references{
Griss et al., bioRxiv 478735; doi: https://doi.org/10.1101/478735
}
\author{
Johannes Griss \email{johannes.griss@meduniwien.ac.at}
}
\keyword{datasets}
