% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{griss_melanoma_rnaseq}
\alias{griss_melanoma_rnaseq}
\title{Raw RNA-seq read counts from the B-Cell
melanoma induction study by Griss et al.}
\format{An object of class \code{DGEList} with 58237 rows and 16 columns.}
\usage{
griss_melanoma_rnaseq
}
\description{
The data is available as a \code{\link[edgeR]{DGEList}} object containing
read counts per gene.
}
\references{
Griss et al., Nat Commun. 2019 10(1):4186. doi: 10.1038/s41467-019-12160-2
}
\author{
Johannes Griss \email{johannes.griss@meduniwien.ac.at}
}
\keyword{datasets}
