% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{jerby_b_cells}
\alias{jerby_b_cells}
\title{Example Seurat object containing B cells
extracted from the single-cell RNA-seq
dataset published by Jerby-Arnon et al.}
\format{An object of class \code{Seurat} with 23686 rows and 920 columns.}
\usage{
jerby_b_cells
}
\description{
This result is stored as a \code{Seurat} object.
}
\references{
Jerby-Arnon et al., Cell 2018 1;175(4):984-997.e24. doi:10.1016/j.cell.2018.09.006
}
\keyword{datasets}
