% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{griss_melanoma_proteomics}
\alias{griss_melanoma_proteomics}
\title{Proteomics intensity-based quantitation data from the B-Cell
melanoma induction study by Griss et al.}
\format{An object of class \code{EList} with 6456 rows and 20 columns.}
\usage{
griss_melanoma_proteomics
}
\description{
The data is available as a \code{\link[limma]{EList}} object containing
the aggregated protein intensity values. Normalisation was already performed
on the PSM level prior to protein-level aggregation.
}
\references{
Griss et al., Nat Commun. 2019 10(1):4186. doi: 10.1038/s41467-019-12160-2
}
\author{
Johannes Griss \email{johannes.griss@meduniwien.ac.at}
}
\keyword{datasets}
