% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{griss_melanoma_result}
\alias{griss_melanoma_result}
\title{Example Camera result created based on the
melanoma induction study by Griss et al.}
\format{An object of class \code{ReactomeAnalysisResult} of length 1.}
\usage{
griss_melanoma_result
}
\description{
The result is stored as a \code{\link[ReactomeGSA]{ReactomeAnalysisResult-class}} object.
}
\references{
Griss et al., Nat Commun. 2019 10(1):4186. doi: 10.1038/s41467-019-12160-2
}
\author{
Johannes Griss \email{johannes.griss@meduniwien.ac.at}
}
\keyword{datasets}
