% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{griss_melanoma_rnaseq}
\alias{griss_melanoma_rnaseq}
\title{Raw RNA-seq read counts from the B-Cell
melanoma induction study by Griss et al.}
\format{An object of class \code{DGEList} with 58237 rows and 16 columns.}
\usage{
griss_melanoma_rnaseq
}
\description{
The data is available as a \code{\link[edgeR]{DGEList}} object containing
read counts per gene.
}
\references{
Griss et al., bioRxiv 478735; doi: https://doi.org/10.1101/478735
}
\author{
Johannes Griss \email{johannes.griss@meduniwien.ac.at}
}
\keyword{datasets}
